refactor: 내 일기 조회 test 분리
